requirements.txt
flask>=3.0,<4.0
flask-login>=0.6,<1.0
flask-wtf>=1.2,<2.0
flask-sqlalchemy>=3.1,<4.0
bcrypt>=4.1,<5.0
python-dotenv>=1.0,<2.0
bleach>=6.1,<7.0
